Create sphere and learn some shortcuts
Created a 3D game object sphere. Learned how to rotate around a game object in the scene view using alt + left click and drag. alt + right click and drag zooms in and out on an object. QWERTY keys are shortcuts for tools like grab, move, rotate, scale. Clicking on the game object in the hierarchy and pressing the f key moves the scene view to that game object and focuses on that.