Fix CRLF to LF error and Learn about Start method
CRLF to LF error was fixed after changing from Visual Studio 2019 to 2022. I also learned about how the start method runs when the game object the script is attached to is in the scene hierarchy and is active and enabled. Code in this method will run before the Update function code and it only runs one time.