Learn how to find x and y rotation
Using the deltaMouse vector, I learned how to get the x and y values for the rotation the object should do. rotationX needs to use mouseDelta.y because moving up and down will rotate the object along the x-axis and vice versa. for mouseDelta.x in rotationY however, it needs to be a negative of that value because mouseDelta.x without the negative would be rotating relative to itself. From our perspective from the camera, that rotation would look like it was moving in the opposite direction then we think it would. Therefore, you need to invert it so it looks like the object is moving in the correct direction for us. For mouseDelta.y, it didn't need to be a negative because from our perspective and from the objects perspective, up and down are the same. If the object was upside down relative to us, then we would have had to invert it.