Assign current mouse pos to prev after rotate
After rotating, keep track of the previous mouse position because the get mouse button down for the other function that would have set the current mouse position to the previous only works when the mouse is first clicked. If the mouse button is held down then it won't make that if statement true again. Therefore, it needs to be set at the end of the rotation so the next frame will have the previous mouse position updated