Change script value in inspector
Learned that if you change a value in the inspector while the game is playing, then it won't save it when you stop playing. Can change values in the inspector instead of having to change them in the script. While playing, when you change the values, it will reflect those changes immediately. This is good for minor changes and play testing because the changes will go away.