Create a Prefab
Learned about how prefabs are created by dragging and dropping a game object from the scene hierarchy to the project window. This saves how the game objects looks at that moment and it can be used to create many more instances of that specific game object. If you change how the game object looks after creating the prefab, it will not change how the prefab looks unless you go in the inspector of the game object that is a prefab and go into overrides. If you do that then all of the instances of the prefab and the prefab in the project window will change to the new edited version of the prefab.