Create Game manager object and gold display
Create empty game object called GM that will hold the game manager script. Create new TextMeshPro element on the Canvas and adjusted the position to the top left. Then dragged and dropped this into the GM game manager script gold display and did the same with the planet game object and set the value of gold to 100.