Initialize variables for enviPiece placement
isSelected will only be true once because each instance of an enviPiece will only be placed on the planet once. mainCam will be used as the starting point of the ray and then the mouse position will be the ending point. objectToPlace will hold the position and rotation of what the placed enviPiece will be