Use Raycasting method
This method shoots the ray and stores information about the object it hit into hitInfo. This is because the keyword out was used. This is used to make the value pass by reference. Without this here, the hitInfo wouldn't store the new information of the object it hit. This information will be used later to know the location of where to put the enviPiece and what rotation it should be. The Physics.Racast function will return a boolean value. If the raycast hit a game object, then it will return true, if not then it will leave the value hitInfo had alone and not run the code inside of the if statement.